requirements.txt
pandas
scikit-learn
joblib
confluent-kafka
upstash-redis
streamlit
altair
python-dotenv